pyproject.toml
[project]
name = "work-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.11",
    "fastapi>=0.128.0",
    "huggingface-hub>=1.3.7",
    "jieba>=0.42.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "loguru>=0.7.3",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "numba>=0.63.0",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pymilvus>=2.6.8",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "regex>=2026.1.15",
    "requests>=2.32.5",
    "types-pyyaml>=6.0.12.20250915",
    "unstructured[pdf]==0.18.31",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = ["mypy>=1.19.1", "pre-commit>=4.5.1", "pytest>=9.0.2", "ruff>=0.14.14"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/silver_pilot"]


# --- Ruff 配置 (Linter & Formatter) ---
[tool.ruff]
# 目标 Python 版本
target-version = "py311"
# 行长限制（大厂通常放宽到 100 或 120，避免过多换行）
line-length = 100

[tool.ruff.lint]
# 启用规则组合：
# E: pycodestyle 错误
# F: Pyflakes (检测未使用的导入/变量)
# I: isort (自动排序 import)
# B: flake8-bugbear (发现潜在 Bug)
# UP: pyupgrade (自动升级旧语法，如 list() -> [])
select = ["E", "F", "I", "B", "UP"]
# 忽略的规则 (如有特殊需要可添加)
ignore = []

[tool.ruff.format]
# 类似于 Black 的格式化风格
quote-style = "double"
indent-style = "space"

# --- Mypy 配置 (静态类型检查) ---
[tool.mypy]
# 严格模式：工业级项目的标配，强制写 Type Hint
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true  # 函数必须有类型标注
ignore_missing_imports = true # 忽略部分没有 stubs 的第三方库


# 针对 src 目录进行检查
files = ["src"]
